grpc-client/requirements.txt
grpcio==1.74.0
protobuf==6.31.1
PyYAML==6.0.1
finnhub-python>=1.3.0
python-dotenv>=1.0.0
